+++ ts/readStream.ts
@@ -19,26 +19,9 @@
   https://www.streampunk.media/ mailto:furnace@streampunk.media
   14 Ormiscaig, Aultbea, Achnasheen, IV22 2JJ  U.K.
 */
-import { Writable, Readable, Transform } from 'stream';
-import type { ffStats } from './types';
-import { BalanceResult, teeBalancer } from './teeBalancer';
-import { localFrame, parallelBalancer } from './parallelBalancer';
-import { serialBalancer } from './serialBalancer';
+import { Readable } from 'stream';
 import { Demuxer } from './types/Demuxer';
-import { Muxer } from './types/Muxer';
-import { Stream } from './types/Stream';
-import { DecodedFrames } from './types/DecodedFrames';
-import type { Governor } from './types/Governor';
-import { Frame } from './types/Frame';
 import { Packet } from './types/Packet';
-import { Timing, TotalTimeed } from './types/Timing';
-import { BeamstreamChannel, BeamstreamParams, BeamstreamSource, BeamstreamStream, ReadableMuxerStream, WritableDemuxerStream } from './types/Beamstreams';
-import { Filterer, FiltererResult, FilterLink } from './types/Filter';
-import { Timable, Timables } from './types/Timable'
-import { EncodedPackets } from './types/Encoder';
-
-import beamcoder from './beamcoder'
-import frameDicer from './frameDicer';
 
 
 export function readStream(params: { highWaterMark?: number }, demuxer: Demuxer, ms: { end: number }, index: number): Readable {
