--- ts/types/Timable.d.ts
@@ -1,7 +1,0 @@
-import { Timing } from './Timing'
-
-export interface Timable {
-    timings?: { [key: string]: Timing; };
-}
-
-export type Timables<T> = Array<T> & Timable;
